+++ db/db_with_timestamp_compaction_test.cc
@@ -152,7 +152,7 @@ TEST_F(TimestampCompatibleCompactionTest, MultipleSubCompactions) {
         0 /* input_level */, 1 /* output_level */, CompactRangeOptions(),
         nullptr /* begin */, nullptr /* end */, true /* exclusive */,
         true /* disallow_trivial_move */,
-        std::numeric_limits<uint64_t>::max() /* max_file_num_to_ignore */,
+        (size_t)std::numeric_limits<uint64_t>::max() /* max_file_num_to_ignore */,
         "" /*trim_ts*/));
   }
