+++ db/memtable_list_test.cc
@@ -47,9 +47,9 @@ class MemTableListTest : public testing::Test {
 
       ColumnFamilyOptions cf_opt1, cf_opt2;
       cf_opt1.cf_paths.emplace_back(dbname + "_one_1",
-                                    std::numeric_limits<uint64_t>::max());
+                                    (size_t)std::numeric_limits<uint64_t>::max());
       cf_opt2.cf_paths.emplace_back(dbname + "_two_1",
-                                    std::numeric_limits<uint64_t>::max());
+                                    (size_t)std::numeric_limits<uint64_t>::max());
       int sz = static_cast<int>(handles.size());
       handles.resize(sz + 2);
       s = db->CreateColumnFamily(cf_opt1, "one", &handles[1]);
@@ -212,7 +212,7 @@ TEST_F(MemTableListTest, Empty) {
 
   autovector<MemTable*> mems;
   list.PickMemtablesToFlush(
-      std::numeric_limits<uint64_t>::max() /* memtable_id */, &mems);
+      (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */, &mems);
   ASSERT_EQ(0, mems.size());
 
   autovector<MemTable*> to_delete;
@@ -441,7 +441,7 @@ TEST_F(MemTableListTest, GetFromHistoryTest) {
   // (It will then be a part of the memtable history).
   autovector<MemTable*> to_flush;
   list.PickMemtablesToFlush(
-      std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
+      (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
   ASSERT_EQ(1, to_flush.size());
 
   MutableCFOptions mutable_cf_options(options);
@@ -503,7 +503,7 @@ TEST_F(MemTableListTest, GetFromHistoryTest) {
 
   to_flush.clear();
   list.PickMemtablesToFlush(
-      std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
+      (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
   ASSERT_EQ(1, to_flush.size());
 
   // Flush second memtable
@@ -634,7 +634,7 @@ TEST_F(MemTableListTest, FlushPendingTest) {
   ASSERT_FALSE(list.imm_flush_needed.load(std::memory_order_acquire));
   autovector<MemTable*> to_flush;
   list.PickMemtablesToFlush(
-      std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
+      (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
   ASSERT_EQ(0, to_flush.size());
 
   // Request a flush even though there is nothing to flush
@@ -644,7 +644,7 @@ TEST_F(MemTableListTest, FlushPendingTest) {
 
   // Attempt to 'flush' to clear request for flush
   list.PickMemtablesToFlush(
-      std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
+      (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
   ASSERT_EQ(0, to_flush.size());
   ASSERT_FALSE(list.IsFlushPending());
   ASSERT_FALSE(list.imm_flush_needed.load(std::memory_order_acquire));
@@ -669,7 +669,7 @@ TEST_F(MemTableListTest, FlushPendingTest) {
 
   // Pick tables to flush
   list.PickMemtablesToFlush(
-      std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
+      (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
   ASSERT_EQ(2, to_flush.size());
   ASSERT_EQ(2, list.NumNotFlushed());
   ASSERT_FALSE(list.IsFlushPending());
@@ -691,7 +691,7 @@ TEST_F(MemTableListTest, FlushPendingTest) {
 
   // Pick tables to flush
   list.PickMemtablesToFlush(
-      std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
+      (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
   ASSERT_EQ(3, to_flush.size());
   ASSERT_EQ(3, list.NumNotFlushed());
   ASSERT_FALSE(list.IsFlushPending());
@@ -700,7 +700,7 @@ TEST_F(MemTableListTest, FlushPendingTest) {
   // Pick tables to flush again
   autovector<MemTable*> to_flush2;
   list.PickMemtablesToFlush(
-      std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush2);
+      (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush2);
   ASSERT_EQ(0, to_flush2.size());
   ASSERT_EQ(3, list.NumNotFlushed());
   ASSERT_FALSE(list.IsFlushPending());
@@ -719,7 +719,7 @@ TEST_F(MemTableListTest, FlushPendingTest) {
 
   // Pick tables to flush again
   list.PickMemtablesToFlush(
-      std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush2);
+      (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush2);
   ASSERT_EQ(1, to_flush2.size());
   ASSERT_EQ(4, list.NumNotFlushed());
   ASSERT_FALSE(list.IsFlushPending());
@@ -741,7 +741,7 @@ TEST_F(MemTableListTest, FlushPendingTest) {
 
   // Pick tables to flush
   list.PickMemtablesToFlush(
-      std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
+      (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush);
   // Should pick 4 of 5 since 1 table has been picked in to_flush2
   ASSERT_EQ(4, to_flush.size());
   ASSERT_EQ(5, list.NumNotFlushed());
@@ -751,7 +751,7 @@ TEST_F(MemTableListTest, FlushPendingTest) {
   // Pick tables to flush again
   autovector<MemTable*> to_flush3;
   list.PickMemtablesToFlush(
-      std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush3);
+      (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */, &to_flush3);
   ASSERT_EQ(0, to_flush3.size());  // nothing not in progress of being flushed
   ASSERT_EQ(5, list.NumNotFlushed());
   ASSERT_FALSE(list.IsFlushPending());
@@ -921,7 +921,7 @@ TEST_F(MemTableListTest, AtomicFlusTest) {
     ASSERT_FALSE(list->IsFlushPending());
     ASSERT_FALSE(list->imm_flush_needed.load(std::memory_order_acquire));
     list->PickMemtablesToFlush(
-        std::numeric_limits<uint64_t>::max() /* memtable_id */,
+        (size_t)std::numeric_limits<uint64_t>::max() /* memtable_id */,
         &flush_candidates[i]);
     ASSERT_EQ(0, flush_candidates[i].size());
   }
